refactor!(components): correct component names
Rename `Header` to `Navbar` and `Hero` to `Header` according to Figma file
