+++ src/app/[locale]/about-us/page.tsx
@@ -1,11 +1,11 @@
-import { Heading, Hero } from "@/components/ui";
+import { Header, Heading } from "@/components/ui";
 
 export default function AboutUs() {
   return (
     <main>
-      <Hero>
+      <Header>
         <Heading level={1}>About us page</Heading>
-      </Hero>
+      </Header>
     </main>
   );
 }
