+++ src/app/[locale]/blog/page.tsx
@@ -1,11 +1,11 @@
-import { Heading, Hero } from "@/components/ui";
+import { Header, Heading } from "@/components/ui";
 
 export default function Blog() {
   return (
     <main>
-      <Hero>
+      <Header>
         <Heading level={1}>Blog page</Heading>
-      </Hero>
+      </Header>
     </main>
   );
 }
