+++ src/app/[locale]/contact-us/page.tsx
@@ -1,11 +1,11 @@
-import { Heading, Hero } from "@/components/ui";
+import { Header, Heading } from "@/components/ui";
 
 export default function ContactUs() {
   return (
     <main>
-      <Hero>
+      <Header>
         <Heading level={1}>Contact us page</Heading>
-      </Hero>
+      </Header>
     </main>
   );
 }
