--- src/components/LayoutHeader/interfaces.ts
@@ -1,1 +1,0 @@
-export type LayoutHeaderProps = PageLocaleParams["params"];
