+++ src/components/LayoutNavbar/interfaces.ts
@@ -1,0 +1,1 @@
+export type LayoutNavbarProps = PageLocaleParams["params"];
