+++ src/components/ui/Header/Header.module.scss
@@ -1,23 +1,10 @@
 .header {
-    background-color: $black;
-    color: $white;
-    padding-top: 18px;
-    padding-bottom: 18px;
-    padding-left: 56px;
-    padding-right: 32px;
+    width: 100%;
     display: flex;
+    flex-direction: column;
+    gap: 4px;
+    justify-content: center;
     align-items: center;
-    justify-content: space-between;
-}
-
-.header__controls {
-    display: flex;
-    align-items: center;
-    gap: 32px;
-}
-
-.header__links {
-    display: flex;
-    align-items: center;
-    gap: 24px;
+    background-color: $lavender;
+    padding: 80px 0px;
 }
