+++ src/components/ui/Header/Header.tsx
@@ -1,20 +1,10 @@
 import styles from "./Header.module.scss";
 import { type HeaderProps } from "./interfaces";
 
-export default function Header({
-  titleSlot,
-  linksSlot,
-  controlsSlot,
-}: HeaderProps) {
+export default function Header({ children }: HeaderProps) {
   return (
-    <header data-testid="header-wrapper" className={styles.header}>
-      {titleSlot}
-      <div data-testid="header-controls" className={styles.header__controls}>
-        <div data-testid="header-links" className={styles.header__links}>
-          {linksSlot}
-        </div>
-        {controlsSlot}
-      </div>
-    </header>
+    <div data-testid="header-wrapper" className={styles.header}>
+      {children}
+    </div>
   );
 }
