--- src/components/ui/Hero/Hero.module.scss
@@ -1,10 +1,0 @@
-.hero {
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    gap: 4px;
-    justify-content: center;
-    align-items: center;
-    background-color: $lavender;
-    padding: 80px 0px;
-}
