--- src/components/ui/Hero/interfaces.ts
@@ -1,3 +1,0 @@
-import { type PropsWithChildren } from "react";
-
-export type HeroProps = PropsWithChildren;
