+++ src/components/ui/Navbar/Navbar.module.scss
@@ -1,0 +1,23 @@
+.navbar {
+    background-color: $black;
+    color: $white;
+    padding-top: 18px;
+    padding-bottom: 18px;
+    padding-left: 56px;
+    padding-right: 32px;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+}
+
+.navbar__controls {
+    display: flex;
+    align-items: center;
+    gap: 32px;
+}
+
+.navbar__links {
+    display: flex;
+    align-items: center;
+    gap: 24px;
+}
