+++ src/components/ui/Navbar/interfaces.ts
@@ -1,0 +1,7 @@
+import type { ReactNode } from "react";
+
+export interface NavbarProps {
+  titleSlot: ReactNode;
+  linksSlot: ReactNode;
+  controlsSlot: ReactNode;
+}
