+++ src/components/ui/index.ts
@@ -3,5 +3,5 @@ export { default as Button } from "./Button/Button";
 export { default as Display } from "./Display/Display";
 export { default as Header } from "./Header/Header";
 export { default as Heading } from "./Heading/Heading";
-export { default as Hero } from "./Hero/Hero";
 export { default as Label } from "./Label/Label";
+export { default as Navbar } from "./Navbar/Navbar";
