--- src/constants/header.ts
@@ -1,1 +1,0 @@
-export const headerVideoLink = "https://www.youtube.com/watch?v=dQw4w9WgXcQ";
