+++ src/constants/navbar.ts
@@ -1,0 +1,1 @@
+export const navbarVideoLink = "https://www.youtube.com/watch?v=dQw4w9WgXcQ";
