+++ src/locale/en.json
@@ -5,7 +5,7 @@
         "heading": "Privacy Policy",
         "body": "Last Updated on 27th January 2022"
     },
-    "header": {
+    "navbar": {
         "title": "Modsen Client Blog",
         "links": {
             "home": "Home",
