+++ src/locale/ru.json
@@ -5,7 +5,7 @@
         "heading": "Политика приватности",
         "body": "Последнее обновление: 27 января 2022"
     },
-    "header": {
+    "navbar": {
         "title": "Блог клиента Modsen",
         "links": {
             "home": "Главная",
